Fix image-orientation with the slow painting path
The painting path that uses tiled image painting to draw
background images was incorrectly orientating images sized
with contains or covers. The problem was in using an
un-oriented image size when correcting the source rect
for orientation, when an oriented size is expected.

In hindsight, we should be using the oriented size for the
tile size, subsetting and other operations in the slow path.
This also simplifies the repeated tile case.

The test fails due to decoding artifacts from the oriented
images, but the reference result also shows bad antialiasing
artifacts around the edges. We need fuzzy matching.

Bug: 1129956
Change-Id: I584b821f110720854fc21df49168afcf37fdc09b