Only do blending for ::selection background for replaced elements
We did a BlendWithWhite() to not have custom selection backgrounds fully
obscure replaced content like images. But, we also did that blending for
text which means the author could not fully control the ::selection
background. Instead, only do this blending for replaced content. This is
in line with what Firefox does and what the spec says[1].

The blending is still done on the UA default background color for both
text and replaced content in LayoutTheme.

[1] https://drafts.csswg.org/css-pseudo-4/#highlight-replaced

Bug: 1018450
Change-Id: Ibb2c63d549a21706a511bbb47ff4c686b246314b